% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add}
\alias{add}
\title{Add together two numbers test for package
how about the second line?}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
A number.
}
\description{
Add together two numbers test for package
how about the second line?
}
\examples{
add(1, 1) #add comments
add(10, 1)
}
