% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{gene.data}
\alias{gene.data}
\title{Simulating data generation, single outcome.}
\usage{
gene.data(
  num_audio,
  patient,
  non_zero_audio,
  abn.audio.mean,
  medium.mean,
  medium_number,
  nor.audio.mean,
  age.mean,
  age.std,
  goodhear.p,
  littletrouble.p,
  coef_age,
  coef_age.sq,
  coef_goodhear,
  coef_littletrouble,
  RMSE,
  seed_idx = 1
)
}
\arguments{
\item{num_audio}{number of audiologists.}

\item{patient}{number of patients evaluated by each audiologist.}

\item{non_zero_audio}{number of significant outlier audiologists.}

\item{abn.audio.mean}{effect of the significant outlier audiologists}

\item{medium.mean}{effect of the intermediate outlier audiologists}

\item{medium_number}{number of intermediate outlier audiologists.}

\item{age.mean}{mean of simulated covariate age for the patients.}

\item{age.std}{std of simulated covariate age for the patients.}

\item{goodhear.p}{probability that hearing status is good.}

\item{littletrouble.p}{probability that hearing status is having a little trouble.}

\item{coef_age}{coefficient of age.}

\item{coef_age.sq}{coefficient of age^2.}

\item{coef_goodhear}{coefficient of hearing status good.}

\item{coef_littletrouble}{coefficient of hearing status having a little trouble.}

\item{RMSE}{standard deviation of random effects.}

\item{seed_idx}{random seed for random effects.}
}
\value{
A dataframe, with number of columns=\emph{4+num_audio},
rows=\emph{patient}*\emph{num_audio}.
}
\description{
Generate the Data which is similar with the CHEARS AAA data.
The predictors are \emph{age}, \emph{age}^2, \emph{hearing status}
(3 levels, good, a little trouble, and bad). \emph{age} is generated from a
normal distribution, mean is \emph{age.mean} and
 the standard deviation is \emph{age.std}.
\emph{Hearing status} is generated from multinormial distribution, with
probabilities \emph{goodhear.p}, and \emph{littletrouble.p}.
}
\examples{
add(1, 1) #add comments
add(10, 1)
}
